+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - ASSET response was added, switches handling from PAGE[404].  Theme assets are also accessible when maintenance mode is turned on.
 - Improvements to session handling
 - Parser class was removed in favor of zbateson/MailMimeParser
+- Phroses\JsonOutput and Phroses\JsonOutputError have been removed
 
 <br />